+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fix crashing Blender when running in background mode (e.g. without GUI).
 - Flamenco: Include extra job parameters to allow for encoding a video at the end of a render
   job that produced an image sequence.
+- Flamenco: Compress all blend files, and not just the one we save from Blender.
+  Requires BAT 0.5 or newer.
 
 
 ## Version 1.9.4 (2018-11-01)